change: moved `service` and `upstream` source file to `http` folder.
