+++ lua/apisix/admin/upstreams.lua
@@ -1,5 +1,5 @@
 local core = require("apisix.core")
-local get_routes = require("apisix.route").routes
+local get_routes = require("apisix.http.route").routes
 local get_services = require("apisix.service").services
 local tostring = tostring
 local ipairs = ipairs
